==> 02_pushshift_comment_scraping.ipynb <==
import requests
import time
import pandas as pd

import datetime as dt

def scrape_pushshift(subreddit, content, start_date, end_date, search_range):
    if 24 % search_range != 0:
        print('not a valid search_range')
        return None
    
    # setting up number of inital conditions
    hours = ((end_date-start_date).days + 1) * (24//search_range)
    before = (dt.date.today()-end_date).days*24
    after = before + search_range
    completed_hours = 0
    completed_days = 0
    
    dict_list = []

    for _ in range(hours):
        # sets up url with the specified subreddit and next hour block
        url = f'https://api.pushshift.io/reddit/search/{content}/?subreddit={subreddit}&size=500&before={str(before)}h&after={str(after)}h'
        res = requests.get(url)
        
        # checks for a valid connection
        while res.status_code != 200:
            print(res.status_code)
            res = requests.get(url)
            time.sleep(1)
        
        the_json = res.json()
        dict_list.extend(the_json['data'])
        
        before += search_range
        after += search_range
        completed_hours += search_range
        
        if completed_hours % 240 == 0:
            completed_days += 10
            print(f'Days Complete: {completed_days}')
        time.sleep(1)

    return dict_list

def scrape_to_dataframe(data):
    posts = []
    for dct in data:
        post_dict = {}
        post_dict['subreddit'] = dct['subreddit']
        post_dict['body'] = dct['body']
        post_dict['author'] = dct['author']
        post_dict['score'] = dct['score']
        post_dict['date'] = dct["created_utc"]
        posts.append(post_dict)

    df =  pd.DataFrame(posts)
    
    def get_date(created):
            return dt.date.fromtimestamp(created)
    df['date'] = df['date'].apply(get_date)
    
    return df

donald_comments_scrape = scrape_pushshift(subreddit = 'the_donald', 
                                          content = 'comment', 
                                          start_date = dt.date(2018, 11, 1), 
                                          end_date = dt.date(2018, 11, 30),
                                          search_range = 12)

donald_comments_data = scrape_to_dataframe(donald_comments_scrape)

donald_comments_data.head()

chapo_comments_scrape = scrape_pushshift(subreddit = 'chapotraphouse', 
                                          content = 'comment', 
                                          start_date = dt.date(2018, 11, 1), 
                                          end_date = dt.date(2018, 11, 30),
                                          search_range = 12)

chapo_comments_data = scrape_to_dataframe(chapo_comments_scrape)
chapo_comments_data.head()

combined_data = pd.concat([donald_comments_data, chapo_comments_data], axis=0, ignore_index=True)

combined_data.body.value_counts()

reminder_mask = combined_data.body.map(lambda x: True if 'Reminder: when someone links to a thread outside of /r/ChapoTrapHouse' in x else False)
combined_data = combined_data.loc[~reminder_mask]

newsnetwork_mask = combined_data.body.map(lambda x: True if 'f8SJIS3XxU8iqHoInw' in x else False)
combined_data = combined_data.loc[~newsnetwork_mask]

automatic_mask = combined_data.body.map(lambda x: True if 'Your comment was automatically removed' in x else False)
combined_data = combined_data.loc[~automatic_mask]

snapshillbot_mask = combined_data.body.map(lambda x: True if 'SnapshillBot' in x else False)
combined_data = combined_data.loc[~snapshillbot_mask]

removal_mask = (combined_data.body == '[removed]') | (combined_data.body == '[deleted]') 
combined_data = combined_data.loc[~removal_mask, :]

combined_data.dropna(inplace=True)
combined_data.reset_index(drop=True, inplace=True)

#combined_data.to_csv('./datasets/raw_donald_chappo_text.csv')

---

Text Cleaning

from bs4 import BeautifulSoup      
from nltk.corpus import stopwords
import regex as re

from nltk.stem
from sklearn.feature_extraction.text import CountVectorizer, HashingVectorizer, TfidfVectorizer

def clean_text(raw_text):
    text = re.sub(r'https:[^\s]+', repl='', string=raw_text)
    text = re.sub(r'http:[^\s]+', repl='', string=text)
    text = BeautifulSoup(text, 'lxml').get_text() 
    text = text.lower()
    words_only = re.sub("[^a-zA-Z]", " ", text)
    return(words_only)

Both the WordLemmeatier and PorterStemmer from NLTK were tried, however the unaltered text consistently performed better. 

combined_data.body = combined_data.body.apply(clean_text)

combined_data.head()

combined_data.shape

#combined_data.to_csv('./datasets/cleaned_donald_chappo_text.csv')

---

# modifying the function to include a comment subject filter, and changing tracker to 100 days

def scrape_pushshift(subreddit, content, q, start_date, end_date, search_range):
    if 24 % search_range != 0:
        print('not a valid search_range')
        return None
    
    # setting up number of inital conditions
    hours = ((end_date-start_date).days + 1) * (24//search_range)
    before = (dt.date.today()-end_date).days*24
    after = before + search_range
    completed_hours = 0
    completed_days = 0
    
    dict_list = []

    for _ in range(hours):
        # sets up url with the specified subreddit and next hour block
        url = f'https://api.pushshift.io/reddit/search/{content}/?subreddit={subreddit}&size=500&before={str(before)}h&after={str(after)}h&q={q}'
        res = requests.get(url)
        
        # checks for a valid connection
        while res.status_code != 200:
            print(res.status_code)
            res = requests.get(url)
            time.sleep(1)
        
        the_json = res.json()
        dict_list.extend(the_json['data'])
        
        before += search_range
        after += search_range
        completed_hours += search_range
        
        if completed_hours % 2400 == 0:
            completed_days += 100
            print(f'Days Complete: {completed_days}')
        time.sleep(1)

    return dict_list

mccain_scrapes = scrape_pushshift(subreddit = 'the_donald', 
                                          content = 'comment',
                                          q = 'McCain',
                                          start_date = dt.date(2015, 1, 1), 
                                          end_date = dt.date.today(),
                                          search_range = 24)

mccain_data = scrape_to_dataframe(mccain_scrapes)
mccain_data.shape

mccain_data.head()

mccain_data.to_csv('./datasets/mccain_data.csv', index=False)

orange_man_bad_scrapes = scrape_pushshift(subreddit = 'the_donald', 
                                          content = 'comment',
                                          q = 'orange man bad',
                                          start_date = dt.date(2015, 1, 1), 
                                          end_date = dt.date.today(),
                                          search_range = 24)

orange_man_bad_data = scrape_to_dataframe(orange_man_bad_scrapes)
orange_man_bad_data.shape

orange_man_bad_data.head()

orange_man_bad_data.to_csv('./datasets/orange_man_bad_data.csv', index=False)

god_emperor_scrapes = scrape_pushshift(subreddit = 'the_donald', 
                                          content = 'comment',
                                          q = 'god emperor',
                                          start_date = dt.date(2015, 1, 1), 
                                          end_date = dt.date.today(),
                                          search_range = 24)

god_emperor_data = scrape_to_dataframe(god_emperor_scrapes)
god_emperor_data.shape

god_emperor_data.head()

god_emperor_data.to_csv('./datasets/god_emperor_data.csv', index=False)

build_the_wall_scrapes = scrape_pushshift(subreddit = 'the_donald', 
                                          content = 'comment',
                                          q = 'build the wall',
                                          start_date = dt.date(2015, 1, 1), 
                                          end_date = dt.date.today(),
                                          search_range = 24)

build_the_wall_data = scrape_to_dataframe(build_the_wall_scrapes)
build_the_wall_data.shape

build_the_wall_data.head()

build_the_wall_data.to_csv('./datasets/build_the_wall.csv', index=False)
