==> 03_stacked_model.ipynb <==
Given that interpretation has been conducted with the count vectorizer and logistic regression, this notebook will focus solely on optimizing predictive performance by incorporating multiple classification models and combining the outputs of these models together. 

# importing the libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer, HashingVectorizer, TfidfVectorizer

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.svm import SVC
from xgboost import XGBClassifier
from sklearn.metrics import confusion_matrix, accuracy_score

from sklearn.model_selection import GridSearchCV
from sklearn.pipeline import Pipeline

train_text = pd.read_csv('./datasets/cleaned_donald_chappo_text.csv', index_col=0)
train_text.dropna(inplace=True)
train_text.head()

train_text.shape

X = train_text['body']
y = train_text['subreddit'].map(lambda x: 1 if x == 'The_Donald' else 0)

# baseline score is approximately 51%
y.value_counts(normalize=True)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = .05, stratify=y)

---

#building a function that builds a pipeline with the specified vectorizer and model for efficient grid searching

def vect_model(vectorizer, model):    
    vect_dict = {
        'count' : CountVectorizer(), 
        'hash' : HashingVectorizer(non_negative=True),
        'td' : TfidfVectorizer()
    }

    model_dict = {
        'log_reg' : LogisticRegression(), 
        'rand_forest' : RandomForestClassifier(),
        'bayes' : MultinomialNB(), 
        'xgb' : XGBClassifier()
    }

    steps = [
        ('vectorize', vect_dict[vectorizer]),
        ('model', model_dict[model])
    ]

    return Pipeline(steps = steps)

Testing different vectorizors with Naive Bayes

bayes_count_model = vect_model('count', 'bayes')

pipe_params = {
    'vectorize__stop_words': ['english', None],
    'vectorize__ngram_range': [(1,1), (1,2)],
    'vectorize__max_features' : [35000, 40000]
}

grid = GridSearchCV(bayes_count_model, pipe_params, n_jobs=-2, verbose=1)
grid.fit(X_train, y_train)

grid.best_params_

grid.best_estimator_.score(X_train, y_train)

grid.best_estimator_.score(X_test, y_test)

bayes_model = grid.best_estimator_

bayes_td_model = vect_model('td', 'bayes')

pipe_params = {
    'vectorize__stop_words': ['english', None],
    'vectorize__ngram_range': [(1,1), (1,2)],
    'vectorize__max_features' : [35000, 40000]
}

grid = GridSearchCV(bayes_td_model, pipe_params, n_jobs=-2, verbose=1)
grid.fit(X_train, y_train)

grid.best_params_

grid.best_estimator_.score(X_train, y_train)

grid.best_estimator_.score(X_test, y_test)

if grid.best_estimator_.score(X_test, y_test) > bayes_model.score(X_test, y_test):
    bayes_model = grid.best_estimator_
    print('model changed')

The score is identical, so the count vercorizer model is kept as native bayes model

---

Testing different vectorizors with Logistic Regression

log_count_model = vect_model('count', 'log_reg')

pipe_params = {
    'vectorize__stop_words': ['english', None],
    'vectorize__ngram_range': [(1,1), (1,2), (1,3)],
    'vectorize__max_features' : [40000],
    'model__penalty' : ['l1', 'l2'],
    'model__C' : [1, .1]
}

grid = GridSearchCV(log_count_model, pipe_params, n_jobs=-2, verbose=1)
grid.fit(X_train, y_train)

grid.best_params_

grid.best_estimator_.score(X_train, y_train)

grid.best_estimator_.score(X_test, y_test)

log_reg_model = grid.best_estimator_

log_td_model = vect_model('td', 'log_reg')

pipe_params = {
    'vectorize__stop_words': ['english', None],
    'vectorize__ngram_range': [(1,1), (1,2)],
    'vectorize__max_features' : [35000, 40000],
}

grid = GridSearchCV(log_td_model, pipe_params, n_jobs=-2, verbose=1)
grid.fit(X_train, y_train)

grid.best_params_

grid.best_estimator_.score(X_train, y_train)

grid.best_estimator_.score(X_test, y_test)

if grid.best_estimator_.score(X_test, y_test) > log_reg_model.score(X_test, y_test):
    log_reg_model = grid.best_estimator_
    print('model changed')

The count vectorizer outperformed the TFIDF vectorizer for logistic regression

---

Testing different vectorizors with XGBoost

import warnings
warnings.filterwarnings("ignore")

xgb_count_model = vect_model('count', 'xgb')

pipe_params = {
    'vectorize__stop_words': [None],
    'vectorize__ngram_range': [(1,1), (1,2)],
    'vectorize__max_features' : [35000, 40000],
    'model__n_estimators' : [300],
    'model__max_depth' : [10]
}

grid = GridSearchCV(xgb_count_model, pipe_params, n_jobs=-2, verbose=1)
grid.fit(X_train, y_train)

grid.best_params_

grid.best_estimator_.score(X_train, y_train)

grid.best_estimator_.score(X_test, y_test)

xgb_model = grid.best_estimator_

xgb_td_model = vect_model('td', 'xgb')

pipe_params = {
    'vectorize__stop_words': ['english'],
    'vectorize__ngram_range': [(1,1),(1,2)],
    'vectorize__max_features' : [35000, 40000],
    'model__n_estimators' : [300],
    'model__max_depth' : [10]
}

grid = GridSearchCV(xgb_td_model, pipe_params, n_jobs=-2, verbose=1)
grid.fit(X_train, y_train)

grid.best_params_

grid.best_estimator_.score(X_train, y_train)

grid.best_estimator_.score(X_test, y_test)

if grid.best_estimator_.score(X_test, y_test) > xgb_model.score(X_test, y_test):
    xgb_model = grid.best_estimator_
    print('model changed')

The count vectorizer outperformed the TFIDF vectorizer for XGBoost

---

Testing different models with Random Forest

rf_count_model = vect_model('count', 'rand_forest')

pipe_params = {
    'vectorize__stop_words': [None],
    'vectorize__ngram_range': [(1,1),(1,2)],
    'vectorize__max_features' : [35000, 40000],
     'model__n_estimators' : [405],
     'model__max_depth': [25]
}

grid = GridSearchCV(rf_count_model, pipe_params, verbose=1)
grid.fit(X_train, y_train)

grid.best_params_

grid.best_estimator_.score(X_train, y_train)

grid.best_estimator_.score(X_test, y_test)

rf_model = grid.best_estimator_

rf_td_model = vect_model('td', 'rand_forest')

pipe_params = {
    'vectorize__stop_words': ['english'],
    'vectorize__ngram_range': [(1,1), (1,2)],
    'vectorize__max_features' : [35000, 40000],
    'model__n_estimators' : [410],
    'model__max_depth': [27]
}

grid = GridSearchCV(rf_td_model, pipe_params, n_jobs=-2, verbose=1)
grid.fit(X_train, y_train)

grid.best_params_

grid.best_estimator_.score(X_train, y_train)

grid.best_estimator_.score(X_test, y_test)

if grid.best_estimator_.score(X_test, y_test) > rf_model.score(X_test, y_test):
    rf_model = grid.best_estimator_
    print('model changed')

The count vectorizer outperformed the TFIDF vectorizer for Random Forest

---

While these accuracies are not unexpected given the similiarity of our subreddits, perhaps they can be improved by looking at the outputs of all our models together. 

# Creating a dataframe that takes the predictions from each model as new training features and counts the votes
predictions_df = pd.DataFrame(index=X_test.index, columns=['bayes', 'log_reg', 'xgb', 'rand_forest'])
predictions_df['bayes'] = bayes_model.predict(X_test)
predictions_df['log_reg'] = log_reg_model.predict(X_test)
predictions_df['xgb'] = xgb_model.predict(X_test)
predictions_df['rand_forest'] = rf_model.predict(X_test)

predictions_df['votes'] = predictions_df.apply(np.sum, axis=1)

predictions_df.head()

# Counting any comment where at least 3 models voted postitve as a positive
y_pred = predictions_df.votes.map(lambda x: 1 if x >= 3 else 0)
accuracy_score(y_test, y_pred)

Slight improvement, but certainly not significant enough to be worth the additional effort. Going forward I would use the Naive Bayes model.

---

For the sake of curiosity we can also try a stacked-model and train a logistic regression on the model outputs and see if this improves accuarcy

# First we need to train the model on the training set predictions
train_predictions_df = pd.DataFrame(index=X_train.index, columns=['bayes', 'log_reg', 'xgb', 'rand_forest'])
train_predictions_df['bayes'] = bayes_model.predict(X_train)
train_predictions_df['log_reg'] = log_reg_model.predict(X_train)
train_predictions_df['xgb'] = xgb_model.predict(X_train)
train_predictions_df['rand_forest'] = rf_model.predict(X_train)

train_predictions_df.head()

god_log_reg = LogisticRegression()
god_log_reg.fit(train_predictions_df, y_train)

god_log_reg.score(train_predictions_df, y_train)

Now we can score it on final model predictions.

god_log_reg.score(test_predictions_df.drop(['votes'], axis=1), y_test)

No boost in performance.
