==> 04_bitcoin_vs_zelda_test.ipynb <==
To test the effectiveness of our approach, we will duplicate it on completely separate subreddits to see the results.

# importing the libraries
import requests
import time
import pandas as pd
import datetime as dt

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB

from sklearn.model_selection import GridSearchCV
from sklearn.pipeline import Pipeline

def scrape_pushshift(subreddit, content, start_date, end_date, search_range):
    if 24 % search_range != 0:
        print('not a valid search_range')
        return None
    
    # setting up number of inital conditions
    hours = ((end_date-start_date).days + 1) * (24//search_range)
    before = (dt.date.today()-end_date).days*24
    after = before + search_range
    completed_hours = 0
    completed_days = 0
    
    dict_list = []

    for _ in range(hours):
        # sets up url with the specified subreddit and next hour block
        url = f'https://api.pushshift.io/reddit/search/{content}/?subreddit={subreddit}&size=500&before={str(before)}h&after={str(after)}h'
        res = requests.get(url)
        
        # checks for a valid connection
        while res.status_code != 200:
            print(res.status_code)
            res = requests.get(url)
            time.sleep(1)
        
        the_json = res.json()
        dict_list.extend(the_json['data'])
        
        before += search_range
        after += search_range
        completed_hours += search_range
        
        if completed_hours % 240 == 0:
            completed_days += 10
            print(f'Days Complete: {completed_days}')
        time.sleep(1)

    return dict_list

def scrape_to_dataframe(data):
    posts = []
    for dct in data:
        post_dict = {}
        post_dict['subreddit'] = dct['subreddit']
        post_dict['title'] = dct['title']
        post_dict['author'] = dct['author']
        post_dict['score'] = dct['score']
        post_dict['domain'] = dct['domain']
        post_dict['source_url'] = dct['url']
        posts.append(post_dict)

    return pd.DataFrame(posts)

zelda_scrape = scrape_pushshift(subreddit = 'zelda', 
                                content = 'submission', 
                                start_date = dt.date(2018, 9, 1), 
                                end_date = dt.date(2018, 11, 30),
                                search_range = 24)

zelda_data = scrape_to_dataframe(zelda_scrape)
zelda_data.head()

bitcoin_scrape = scrape_pushshift(subreddit = 'bitcoin', 
                                  content = 'submission', 
                                  start_date = dt.date(2018, 11, 1), 
                                  end_date = dt.date(2018, 11, 30),
                                  search_range = 24)

bitcoin_data = scrape_to_dataframe(bitcoin_scrape)
bitcoin_data.head()

combined_data = pd.concat([zelda_data, bitcoin_data], axis=0, ignore_index=True)

removal_mask = (combined_data.title == '[removed]') | (combined_data.title == '[deleted]') 
combined_data = combined_data.loc[~removal_mask, :]
combined_data.dropna(inplace=True)
combined_data.reset_index(inplace=True)

combined_data.shape

Text Cleaning

from bs4 import BeautifulSoup      
from nltk.corpus import stopwords
import regex as re

def clean_text(raw_text):
    text = re.sub(r'https:[^\s]+', repl='', string=raw_text)
    text = re.sub(r'http:[^\s]+', repl='', string=text)
    text = BeautifulSoup(text, 'lxml').get_text() 
    text = text.lower()
    words_only = re.sub("[^a-zA-Z]", " ", text)
    return(words_only)

combined_data.title = combined_data.title.map(clean_text)

#combined_data.to_csv('./datasets/zelda-bitcoin.csv', index=False, index_label=None)

X = combined_data['title']
y = combined_data['subreddit'].map(lambda x: 1 if x == 'Bitcoin' else 0)

#baseline
y.value_counts(normalize=True)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = .20, stratify=y)

---

def vect_model(vectorizer, model):    
    vect_dict = {
        'count' : CountVectorizer(), 
        'td' : TfidfVectorizer()
    }

    model_dict = {
        'bayes' : MultinomialNB(), 
    }

    steps = [
        ('vectorize', vect_dict[vectorizer]),
        ('model', model_dict[model])
    ]

    return Pipeline(steps = steps)

Testing different vectorizors with Naive Bayes

bayes_count_model = vect_model('count', 'bayes')

pipe_params = {
    'vectorize__stop_words': ['english', None],
    'vectorize__ngram_range': [(1,1), (1,2)],
    'vectorize__max_features' : [5000, 10000, None]
}

grid = GridSearchCV(bayes_count_model, pipe_params, n_jobs=1, verbose=1)
grid.fit(X_train, y_train)

grid.best_params_

grid.best_estimator_.score(X_train, y_train)

grid.best_estimator_.score(X_test, y_test)

---

# testing with tifidf
bayes_td_model = vect_model('td', 'bayes')

pipe_params = {
    'vectorize__stop_words': ['english', None],
    'vectorize__ngram_range': [(1,1), (1,2)],
    'vectorize__max_features' : [5000, 10000, None],
    'vectorize__lowercase' : [True, False],
    'vectorize__norm' : ['l1', 'l2']
}

grid = GridSearchCV(bayes_td_model, pipe_params, n_jobs=-2, verbose=1)
grid.fit(X_train, y_train)

grid.best_params_

grid.best_estimator_.score(X_train, y_train)

grid.best_estimator_.score(X_test, y_test)

Given the high accuarcy on train and test data for both naive bayes and logistic regression, we can conclude that this methodology works well given two subreddits with completly unrelated content.
