==> 05_text_analysis.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction import text

from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB

pd.options.display.max_colwidth = 500

train_text = pd.read_csv('./datasets/cleaned_donald_chappo_text.csv', index_col=0)
train_text.dropna(inplace=True)
train_text.head()

train_text.shape

X = train_text['body']
y = train_text['subreddit'].map(lambda x: 1 if x == 'The_Donald' else 0)

#baseline
y.value_counts(normalize=True)

Analyzing Text with Logistic Regression (no stopwords)

def text_dataframe(text, n_features, sw=None):
    count = CountVectorizer(stop_words=sw, max_features=n_features)
    text = count.fit_transform(text)
    text = text.toarray()
    return pd.DataFrame(text, columns=count.get_feature_names())

X_vect = text_dataframe(X, n_features=40000, sw='english')

logit = LogisticRegression()

logit.fit(X_vect, y)

logit.score(X_vect, y)

Creating a dictonary with key values pairs that correspond to the odds and associated word as calculated by the logistic regressoin

coef_dict = {}

for i in zip(np.exp(logit.coef_[0]),X_vect.columns):
    coef_dict[float(i[0])] = i[1]

# taking the largest 15 odds, these correlate strongly text from The Donald
donald_top_ten = pd.Series(np.exp(logit.coef_[0])).sort_values(ascending=False).head(15)

# taking the smallest 15 odds, these correlate strongly text from ChapoTrapHouse
chapo_top_ten = pd.Series(np.exp(logit.coef_[0])).sort_values(ascending=True).head(15)

The top 15 words that correlate with The Donald and their associated odds

for i in donald_top_ten:
    print(coef_dict[i], '---', i)

The top 15 words that correlate with ChapoTrapHouse and their associated odds (relative to being in The Donald)

for i in chapo_top_ten:
    print(coef_dict[i], '---', i)

Analyzing Text with Logistic Regression (stopwords)

X_vect = text_dataframe(X, n_features=40000)

logit.fit(X_vect, y)

coef_dict = {}

for i in zip(np.exp(logit.coef_[0]),X_vect.columns):
    coef_dict[float(i[0])] = i[1]

donald_top_ten = pd.Series(np.exp(logit.coef_[0])).sort_values(ascending=False).head(15)

chapo_top_ten = pd.Series(np.exp(logit.coef_[0])).sort_values(ascending=True).head(15)

for i in donald_top_ten:
    print(i, '---', coef_dict[i])

for i in chapo_top_ten:
    print(i, '---', coef_dict[i])

---

Analyzing the most frequent words in both subreddits

#breaking up the combined data by subreddit for simplicity
the_donald = train_text.loc[train_text['subreddit'] == 'The_Donald']
chapo = train_text.loc[train_text['subreddit'] == 'ChapoTrapHouse']

snapshillbot_mask = chapo.body.map(lambda x: True if 'SnapshillBot' in x else False)
chapo = chapo.loc[~snapshillbot_mask]

the_donald.shape

chapo.shape

#defining a function to return a dataframe with the total count of each word in the text
def word_frequency(text, sw=None, ngrams=(1,1)):
    count = CountVectorizer(stop_words=sw, ngram_range=ngrams)
    text = count.fit_transform(text)
    sums = text.toarray().sum(axis=0)
    return pd.DataFrame([sums], columns=count.get_feature_names()).T

chapo_no_s_words = word_frequency(chapo.body)

chapo_no_s_words.sort_values(0, ascending=False).head(10)

chapo_s_words = word_frequency(chapo['body'], sw='english')

chapo_s_words.sort_values(0, ascending=False).head(10)

donald_no_s_words = word_frequency(the_donald['body'])

donald_no_s_words.sort_values(0, ascending=False).head(10)

donald_s_words = word_frequency(the_donald['body'], sw='english')

donald_s_words.sort_values(0, ascending=False).head(10)

Calculating the number of unique words in the top 1,000 words for each subreddit

# the 1000 most common words in the donald
donald_top_1000 = list(donald_s_words.sort_values(0, ascending=False).head(1000).index)

# the 1000 most common words in chapo trap house
chapo_top_1000 = list(chapo_s_words.sort_values(0, ascending=False).head(1000).index)

# finding the number of these words that are unique to the donald
donald_specific = []
for word in donald_top_1000:
    if word not in chapo_top_1000:
        donald_specific.append(word)

len(donald_specific)

# finding the number of these words that are unique to the chapo trap house (this should be the same)
donald_specific = []
chapo_specific = []
for word in chapo_top_1000:
    if word not in donald_top_1000:
        chapo_specific.append(word)

len(chapo_specific)

For reference, repeating the process with the Zelda and Bitcoin comments to compare unique word counts

combined_data = pd.read_csv('./datasets/zelda-bitcoin.csv', index_col=0)
combined_data.dropna(inplace=True)
combined_data.shape

zelda = combined_data.loc[combined_data['subreddit'] == 'zelda']

zelda.shape

bitcoin = combined_data.loc[combined_data['subreddit'] == 'Bitcoin']

bitcoin.shape

zelda_words = word_frequency(zelda['title'], sw='english')
bitcoin_words = word_frequency(bitcoin.title, sw='english')

zelda_top_1000 = list(zelda_words.sort_values(0, ascending=False).head(1000).index)
bitcoin_top_1000 = list(bitcoin_words.sort_values(0, ascending=False).head(1000).index)

zelda_specific = []
for word in zelda_top_1000:
    if word not in bitcoin_top_1000:
        zelda_specific.append(word)

len(zelda_specific)

bitcoin_specific = []
for word in bitcoin_top_1000:
    if word not in zelda_top_1000:
        bitcoin_specific.append(word)

len(bitcoin_specific)

Significantly more unique words for the Zelda and Bitcoin subreddits, this makes sense and helps to explain why the classification of these subreddits had a significantly higher accuracy. 

---

# defining a function that returns the shared words out of the n most common words
def common_words(num):
    donald = list(donald_s_words.sort_values(0, ascending=False).head(num).index)
    chapo = list(chapo_s_words.sort_values(0, ascending=False).head(num).index)

    common = []
    for word in donald:
        if word in chapo:
            common.append(word)
            
    return common

# defining a function that returns the unique words for each subreddit out of the n most common words
def seperate_words(text_1, text_2, num):
    text_1 = list(text_1.sort_values(0, ascending=False).head(num).index)
    text_2 = list(text_2.sort_values(0, ascending=False).head(num).index)

    text_1_list = []
    for word in text_1:
        if word not in text_2:
            text_1_list.append(word)
            
    text_2_list = []
    for word in text_2:
        if word not in text_1:
            text_2_list.append(word)
            
    return text_1_list, text_2_list

---

Top phrases in each subreddit

donald_bigrams = word_frequency(the_donald['body'], ngrams=(2,2))
chapo_bigrams = word_frequency(chapo.body, ngrams=(2,2))

unique_donald_bigrams, unique_chapo_bigrams = seperate_words(donald_bigrams, chapo_bigrams, 1000)

unique_donald_bigrams[0:15]

unique_chapo_bigrams[0:15]

donald_trigrams = word_frequency(the_donald['body'], ngrams=(3,3))

chapo_trigrams = word_frequency(chapo.body, ngrams=(3,3))

unique_donald_trigrams, unique_chapo_trigrams = seperate_words(donald_trigrams, chapo_trigrams, 1000)

unique_donald_trigrams[:15]

unique_chapo_trigrams[:15]

# looking at how the phrase orange man bad is used
orange_man_mask = the_donald.body.map(lambda x: True if 'orange man bad' in x else False)
the_donald.loc[orange_man_mask, :].head()

Modeling with custom stop words

common_10 = common_words(10)

stop_words = text.ENGLISH_STOP_WORDS.union(common_10)

X = train_text.body
y = train_text.subreddit.map(lambda x: 1 if x == 'The_Donald' else 0)

count = CountVectorizer(stop_words=common_10)
nb = MultinomialNB()

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = .1, stratify = y, random_state=42)

X_train = count.fit_transform(X_train)
X_test = count.transform(X_test)

X_train.shape

nb.fit(X_train, y_train)
nb.score(X_train, y_train)

nb.score(X_test, y_test)

No significant improvement 

common_100 = common_words(100)

stop_words = text.ENGLISH_STOP_WORDS.union(common_100)

X = train_text.body
y = train_text.subreddit.map(lambda x: 1 if x == 'The_Donald' else 0)

count = CountVectorizer(stop_words=common_100)
nb = MultinomialNB()

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = .1, stratify = y, random_state=42)

X_train = count.fit_transform(X_train)
X_test = count.transform(X_test)

X_train.shape

nb.fit(X_train, y_train)
nb.score(X_train, y_train)

nb.score(X_test, y_test)

Slight Decrease in Performance - Removing common words does not seem to improve the score
